renamed test file and updated documentation
named file into ffmpeg to aac codec because it will take any ffmpeg compatible file and turn it into any audio format that uses aac codec. by sorting scripts as codeces, we reduce redundant scripts. for example, mp3 creation and m4a creation will both use aac codec script. we'll do this by having the script that sends the filename also save the 'ends with' as a variable so we can determine which script to send that whole filename towards using a logic gate before sending. select case for endswith to assign 'codec_used' and that codec_used determimines which script we send it to