created libvorbis converter for '.ogg' encoding, and checked it off the devlog tasklist
ogg won't get used often, but it'll be super useful for those who need the format for their own projects. checked 'ogg' off the devlog checklist. deleted created test files because i don't feel like putting the output folder in the ignoreme because it will be confusing to not have them when I clone this repo.