updated devlog: task 2 is complete
checked off items completed in the previous major input_to_output update. with that, task 2 is complete. we may consider a further update that has input_to_output report failures to the gui, so the gui can in turn both tell the user cleanly, and/or optionally, create a text file report